dev-requirements.txt
pytest
pytest-django
pytest-cov
# optional but pretty
pytest-sugar